pyproject.toml
[project]
name = "mojo-mcp"
version = "0.1.0"
description = "Mojo MCP Server using Code Mode pattern (search + execute)"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
mojo-mcp = "mojo_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mojo_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "mojo: tests that require a working Mojo installation (deselected by default)",
]
addopts = ""
